rename FetchDataHelpers to helpers file, add iterate func from service page
